+++ src/chart/generator/axis.h
@@ -29,7 +29,7 @@ struct ChannelStats
 	template <ChannelIdLike T>
 	[[nodiscard]] const TrackType &at(const T &id) const
 	{
-		return tracked[-id];
+		return tracked[+id];
 	}
 
 	void track(ChannelId at, const Data::MarkerId &id)
@@ -40,13 +40,14 @@ struct ChannelStats
 
 	void track(ChannelId at, const double &value)
 	{
-		std::get<0>(tracked[at]).include(value);
+		if (std::isfinite(value))
+			std::get<0>(tracked[at]).include(value);
 	}
 
 	template <ChannelIdLike Id>
 	void setIfRange(Id at, const Math::Range<> &range)
 	{
-		if (auto *r = std::get_if<0>(&tracked[-at])) *r = range;
+		if (auto *r = std::get_if<0>(&tracked[+at])) *r = range;
 	}
 };