+++ src/chart/generator/marker.cpp
@@ -31,10 +31,10 @@ Marker::Marker(const Options &options,
     bool needMarkerInfo) :
     cellInfo(data.cellInfo(index, needMarkerInfo)),
     mainId(data.getId(mainAxisList,
-        options.dimLabelIndex(-options.mainAxisType()),
+        options.dimLabelIndex(+options.mainAxisType()),
         index)),
     subId(data.getId(subAxisList,
-        options.dimLabelIndex(-options.subAxisType()),
+        options.dimLabelIndex(+options.subAxisType()),
         index)),
     sizeId(data.getId(
         options.getChannels().at(ChannelId::size).dimensions(),
@@ -76,7 +76,7 @@ Marker::Marker(const Options &options,
 	if (subAxisList != options.subAxis().dimensions())
 		subId.label =
 		    data.getId(options.subAxis().dimensions(),
-		            options.dimLabelIndex(-options.subAxisType()),
+		            options.dimLabelIndex(+options.subAxisType()),
 		            index)
 		        .label;
 
@@ -140,6 +140,9 @@ Marker::Marker(const Options &options,
 		        data.joinDimensionValues(labelChannel.dimensions(),
 		            index)};
 	}
+
+	if (!std::isfinite(position.x) || !std::isfinite(position.y))
+		enabled = false;
 }
 
 bool Marker::connectMarkers(bool first,
@@ -220,14 +223,14 @@ void Marker::fromRectangle(const Geom::Rect &rect)
 
 Math::Range<> Marker::getSizeBy(AxisId axisId) const
 {
-	return isHorizontal(+axisId) ? toRectangle().hSize()
-	                             : toRectangle().vSize();
+	return isHorizontal(++axisId) ? toRectangle().hSize()
+	                              : toRectangle().vSize();
 }
 
 void Marker::setSizeBy(AxisId axisId, const Math::Range<> range)
 {
 	auto rect = toRectangle();
-	if (isHorizontal(+axisId))
+	if (isHorizontal(++axisId))
 		rect.setHSize(range);
 	else
 		rect.setVSize(range);