+++ src/chart/options/channels.h
@@ -30,12 +30,12 @@ struct Channels : Refl::EnumArray<ChannelId, Channel>
 	template <ChannelIdLike T>
 	[[nodiscard]] const Channel &at(const T &id) const
 	{
-		return at(-id);
+		return at(+id);
 	}
 
 	template <ChannelIdLike T> [[nodiscard]] Channel &at(const T &id)
 	{
-		return at(-id);
+		return at(+id);
 	}
 
 	[[nodiscard]] const AxisChannelProperties &axisPropsAt(