+++ src/chart/rendering/drawaxes.cpp
@@ -101,7 +101,8 @@ const DrawAxes &&DrawAxes::init() &&
 			        !item.startPos.isAuto() && *item.startPos,
 			        !item.endPos.isAuto() && *item.endPos,
 			        axis.dimension.factor);
-			    needSeparators && sepWeight > 0)
+			    needSeparators && sepWeight > 0
+			    && item.range.getMin() > 0)
 				separators.emplace_back(item.range.getMin(),
 				    sepWeight);
 		}
@@ -144,7 +145,7 @@ void DrawAxes::generateMeasure(Gen::AxisId axisIndex,
     double stepSize,
     double weight)
 {
-	auto orientation = !+axisIndex;
+	auto orientation = !++axisIndex;
 	const auto &meas = getAxis(axisIndex).measure;
 	auto rangeSize = meas.range.size();
 	auto singleLabelRange = Math::Floating::is_zero(rangeSize);
@@ -212,9 +213,9 @@ void DrawAxes::generateMeasure(Gen::AxisId axisIndex,
 
 Geom::Line DrawAxes::getAxisLine(Gen::AxisId axisIndex) const
 {
-	auto offset = this->origo().getCoord(+!axisIndex);
+	auto offset = this->origo().getCoord(!++axisIndex);
 
-	auto direction = Geom::Point::Ident(+axisIndex);
+	auto direction = Geom::Point::Ident(++axisIndex);
 
 	auto p0 = direction.flip() * offset;
 	auto p1 = p0 + direction;
@@ -265,7 +266,9 @@ Geom::Point DrawAxes::getTitleBasePos(Gen::AxisId axisIndex,
 	default:
 	case Pos::min_edge: break;
 	case Pos::max_edge: orthogonal = 1.0; break;
-	case Pos::axis: orthogonal = origo().getCoord(+!axisIndex); break;
+	case Pos::axis:
+		orthogonal = origo().getCoord(!++axisIndex);
+		break;
 	}
 
 	double parallel{0.0};
@@ -439,7 +442,7 @@ void DrawAxes::drawDimensionLabel(Gen::AxisId axisIndex,
     double weight) const
 {
 	if (weight == 0) return;
-	auto orientation = +axisIndex;
+	auto orientation = ++axisIndex;
 
 	const auto &labelStyle = rootStyle.plot.getAxis(axisIndex).label;